handwrite/requirements.txt
paddlepaddle==2.6.1
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
numpy==1.26.4
fastapi==0.114.1
uvicorn==0.30.6
onnx==1.16.2
onnxruntime==1.19.2
pydantic==2.9.2
mlflow==2.16.0
wandb==0.17.9
shapely==2.1.2
scikit-image==0.25.2
imgaug==0.4.0
pyclipper==1.3.0.post6
lmdb==1.7.3
tqdm==4.67.1
visualdl==2.5.3
rapidfuzz==3.14.1
cython==3.1.4
lxml==6.0.2
premailer==3.10.0
openpyxl==3.1.5
attrdict==2.0.1
openai==0.28.1
pillow>=10.1
python-multipart==0.0.6
